requirements.txt
llama-index
pypdf
pinecone-client